+++ packages/xstate-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.1
+
+### Patch Changes
+
+- [`c0bd0407`](https://github.com/davidkpiano/xstate/commit/c0bd040767dcac20ed690e49a8725b4f1011dd5d) [#1493](https://github.com/davidkpiano/xstate/pull/1493) Thanks [@davidkpiano](https://github.com/davidkpiano)! - There will now be a descriptive error when trying to use an actor-like object in the `useService()` hook, where `useActor()` should be preferred:
+
+  > Attempted to use an actor-like object instead of a service in the useService() hook. Please use the useActor() hook instead.
+
 All notable changes to this project will be documented in this file.
 
 ## [1.0.0-rc.7]